* main:
  feat: Gamify authentication with firebase (#279)

+++ libs/gamify/applications/src/lib/applications.controller.ts
@@ -1,22 +1,22 @@
-import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, ParseIntPipe, Post, Put } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, ParseIntPipe, Post, Put, UnauthorizedException } from '@nestjs/common';
 import { ApplicationsService } from './applications.service';
 import { CreateApplicationInput } from './dto/create-application.input';
 import { UpdateApplicationInput } from './dto/update-application.input';
+import { Roles, User, UserModel } from '@gamify/auth';
+import { Role } from '@gamify/core';
 
 @Controller('applications')
 export class ApplicationsController {
   constructor(private applicationsService: ApplicationsService) {}
 
   @Post()
-  async create(@Body() createApplicationInput: CreateApplicationInput) {
-    console.log(JSON.stringify(createApplicationInput));
-
+  @Roles(Role.ADMIN, Role.MODERATOR)
+  async create(@Body() createApplicationInput: CreateApplicationInput, @User() user: UserModel) {
     if (! await this.applicationsService.isNameUnique(createApplicationInput.name)) {
       throw new BadRequestException("Name must be unique");
     }
 
-    // TODO Validate current user id, should be obtained from current authenticated user.
-    createApplicationInput.ownerUserId = 1;
+    createApplicationInput.ownerUserId = user.id;
 
     return await this.applicationsService.create(createApplicationInput);
   }
@@ -32,7 +32,12 @@ export class ApplicationsController {
   }
 
   @Put(':id')
-  async update(@Param('id', ParseIntPipe) id: number, @Body() updateApplicationInput: UpdateApplicationInput, ) {
+  @Roles(Role.ADMIN, Role.MODERATOR)
+  async update(@Param('id', ParseIntPipe) id: number, @Body() updateApplicationInput: UpdateApplicationInput, @User() user: UserModel) {
+    if (! await this.applicationsService.canModerateApplication(id, user.id)) {
+      throw new UnauthorizedException();
+    }
+
     await this.findApplicationOrFail(id);
 
     if (! await this.applicationsService.isNameUnique(updateApplicationInput.name)) {
@@ -43,7 +48,12 @@ export class ApplicationsController {
   }
 
   @Delete(':id')
-  async remove(@Param('id', ParseIntPipe) id: number) {
+  @Roles(Role.ADMIN, Role.MODERATOR)
+  async remove(@Param('id', ParseIntPipe) id: number, @User() user: UserModel) {
+    if (! await this.applicationsService.canModerateApplication(id, user.id)) {
+      throw new UnauthorizedException();
+    }
+
     await this.findApplicationOrFail(id);
 
     return await this.applicationsService.remove(id);
